Move installer and tests to their own files
